Renamed React component
